Prefer double quotes in Ruby
This will reduce the number of changes we have to make: when a string
transitions into something that needs interpolation, we do not need to
change the quotes.

This will also bring consistency to our writing: all strings will now
have double quotes, regardless of their contents.

It also puts us more in line with a large amount of external Ruby code,
which tends to use double quotes.